flow_library_server/pyproject.toml
[tool.poetry]
name = "t-library_server"
version = "0.1.0"
description = ""
authors = ["Denis Tokarev <wf.denis.tokarev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
